Refactor LegalTextVersionAdminForm from admin to forms
